backend/requirements.txt
fastapi==0.128.0
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.12.5
pydantic[email]==2.12.5
email-validator==2.2.0
firebase-admin==6.7.0
openai==1.59.8
pillow==11.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
apscheduler==3.10.4
